src/llm-api/pyproject.toml
[project]
name = "llm-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.125.0",
    "google-cloud-aiplatform>=1.132.0",
    "google-cloud-translate>=3.16.0",
    "google-genai>=0.8.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "uvicorn[standard]>=0.38.0",
]
